requirements.txt
anthropic>=0.25.0
openai>=1.30.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
